refactor: remove search module
