+++ domain-db/src/lib.rs
@@ -4,5 +4,4 @@ extern crate r2d2;
 extern crate r2d2_diesel;
 
 pub mod db;
-pub mod search;
 pub mod sources;
